Refactor shopping-cart
