+++ views/user/shopping-cart.hbs
@@ -18,7 +18,7 @@
             </div>
         </div>
     </section>
-    {{#if noItem}}
+    {{#if noOrder}}
     <section class="section-two">
         <div class="container">
             <div class="section-two-content">
@@ -35,7 +35,6 @@
             </div>
         </div>
     </section>
-    {{/if}}
     <section class="section-three">
         <div class="container">
             <div class="section-three-content">
@@ -111,9 +110,10 @@
                     <a href="/shopping-cart/food-voucher" class="order-button" id="choose-fv">Chọn mã giảm giá</a>
                     <a href="/shopping-cart/ship-voucher" class="order-button" id="choose-fv">Chọn mã shipping</a>
                     <a href="/shopping-cart/address" class="order-button" id="choose-fv">Chọn địa chỉ</a>
-                    <a href="#" class="order-button" id="order-submit">Đặt hàng</a>
+                    <a href="/shopping-cart/submit-order" class="order-button" id="order-submit">Đặt hàng</a>
                 </div>
             </div>
         </div>
     </section>
+    {{/if}}
 </div>
